Add task to install and configure SimpleSAML to work with XDMod  (#3)
* Add task to install and configure SimpleSAML to work with XDMod in clustermon.yml

This commit adds a new task to the `clustermon.yml` playbook to install and configure SimpleSAML on the target machines. The new task uses the `yum` package manager to install the required packages, creates the necessary directories for SimpleSAML, and copies the configuration files to the appropriate locations.

* Extract SimpleSAMLphp archive to /var/www/

This commit combines the tasks of downloading simplesaml from the url and extracting it to destination in one step

* Change task description to 'Download and Extract SimpleSAMLphp'

This commit updates the description of a task to reflect the fact that the task both downloads and extracts the SimpleSAMLphp archive. The new description is more accurate and descriptive than the previous one.

* Add task to create log file

Add a new task to create a log file using the `file` module in Ansible. The task creates a file at the path `/var/log/simplesaml/simplesamlphp.log` with the mode `0664`, owned by the `apache` user and group. This will ensure that the log file is created with the correct permissions and ownership for Apache to write to.

* Download and Extract SimpleSAMLphp

Download SimpleSAMLphp to /tmp dir and extract it to /var/www directory. The version used in this installation is v1.19.7.

Changes made:
- Added task to download SimpleSAMLphp to /tmp dir with get_url module.
- Added task to extract SimpleSAMLphp using unarchive module with src as the downloaded tar.gz file and dest as /var/www/.
- Commented out the command to rename the SimpleSAMLphp directory.

* Restart httpd.service using systemd with elevated privileges. Change service name to httpd from http. Previous service was not responding properly.

* Refactor simplesaml config file to use templates

This commit moves the config.php file for the simplesaml role to a new templates directory. This change allows for the file to be more easily managed and templated for different environments.

This commit also  updates the simplesaml role to use Ansible to copy the config.php file from the roles/simplesaml/templates/ directory to its final destination. This change ensures that the file is properly managed and templated in a repeatable and automated way.

* Remove commented out task and command for SimpleSAMLphp

This commit removes a commented out task and command related to renaming the SimpleSAMLphp directory in the roles/simplesaml/tasks/main.yml file. These lines were previously disabled for testing purposes and are no longer needed in the final version of the role.

* Remove simplesaml config file

This commit removes the roles/simplesaml/files/config.php file from the repository since it has been moved to the roles/simplesaml/templates/ directory.

* feat: Add task to create and clean SimpleSAMLphp directory

This commit adds a task to create the directory /tmp/simplesamlphp/ and delete its contents if it already exists. The task uses the file module with the state=directory option and recurse=yes option to remove all files and directories inside /tmp/simplesamlphp/ before creating it.

The get_url task has also been modified to use a dynamic simplesaml_url variable for the download URL, which allows for greater flexibility.

Finally, the unarchive task has been modified to use lookup plugin with fileglob to search for the archive file in /tmp/simplesamlphp/. The extra_opts parameter is also added with the --strip-components=1 option to remove the first directory component from each file name extracted from the archive. The become parameter is included to ensure that the task runs with elevated privileges.

* feat: Add task to create SimpleSAMLphp log directory

This commit adds a task to create the /var/log/simplesaml/ directory if it doesn't already exist. The file module is used with the state=directory option to create the directory with the mode set to 0775 and owner and group parameters set to apache.

* feat: Add variables and update config file template

This commit adds three variables, trusted_url_domains, session_cookie_domain, and language_cookie_domain, and updates the config.php.j2 file to use these variables. The src field in the template task is also updated from config.php to config.php.j2. This allows for more flexible configuration and makes it easier to update the trusted URL domains and related settings in the future.

This commit also updates the file config.php.j2 to replect appropriate variable changes

* Refactor simplesaml configuration

Update simplesaml configuration files to use non-template config.php
Update config.php with correct destination path
Enable simplesaml authentication with new release v1.19.7
Set trusted_url_domains to 'test.edu' for authentication
This commit modifies the following files:

group_vars/all
roles/simplesaml/tasks/main.yml
roles/simplesaml/templates/config.php
Changes made to roles/simplesaml/tasks/main.yml include using a non-template config.php file and updating the destination path to /etc/xdmod/simplesamlphp/config/config.php. Authentication is enabled with a new release v1.19.7. The trusted_url_domains is set to 'test.edu' for authentication.

Changes made to group_vars/all include setting enable_simplesaml to true and setting trusted_url_domains to ''

* Replace shortened version of simplesaml config.php with original version including comments

The shortened version of the config.php file in roles/simplesaml/templates/ was replaced with the original version which includes detailed comments for better understanding of the configuration

* Added [authentication] section to portal_settings.ini file before the [auto_login] section. Used replace module to modify the file.

* Add [authentication] section to portal_settings.ini

This commit adds a new task to the `simplesaml` role to insert a `[authentication]` section to `portal_settings.ini` configuration file. The new task uses the `blockinfile` module to add the new section on a new line after the `basic_auth = "on"` line. The `marker` parameter is used to add a comment to the file to indicate that the block was added by Ansible and should not be edited manually.

* Update main.yml

* Update authsources.php

---------

Co-authored-by: Krish <krish.94@hotmail.com>
Co-authored-by: root <root@ohpc.openstack.internal>
Co-authored-by: eesaanatluri <39744656+eesaanatluri@users.noreply.github.com>