Add 'openssl_cert_gen' role for OpenSSL certificate generation  (#5)
* Add 'openssl_cert_gen' role for OpenSSL certificate generation and update clustermon.yml and group_vars/all

This commit adds the 'openssl_cert_gen' role to the Ansible project, which provides tasks to generate an OpenSSL private key, create a Certificate Signing Request (CSR), and generate a self-signed OpenSSL certificate using the private key and CSR. This role can be used to automate the process of generating certificates for servers or applications in a secure and scalable manner.

Additionally, this commit updates the 'clustermon.yml' playbook and 'group_vars/all' inventory file to reflect the necessary variables for the 'openssl_cert_gen' role and adds a task to include the 'openssl_cert_gen' role in the main playbook.

* Add Apache HTTPS virtual host configuration

This commit adds an Apache virtual host configuration file for serving XDMod over HTTPS. The configuration file sets up a permanent redirect from HTTP to HTTPS, and includes SSL certificate and key file paths. Additionally, it specifies the document root, directory options, and access control for the application, along with an alias for SimpleSAMLphp. Error and access logging is also configured for the virtual host. This file is intended to be used in conjunction with the existing XDMod Apache configuration.

This commit also adds this template as a part of the openssl_cert_gen role to put the modified xdmod apache file in the system location /etc/httpd/conf.d/

* Update xdmod.conf

* Refactor OpenSSL certificate generation task

This commit refactors the OpenSSL certificate generation task in the `openssl_cert_gen` role.

* Update SSL certificate paths in Apache configuration

This commit updates the SSLCertificateFile and SSLCertificateKeyFile paths in the Apache configuration to use the correct paths for the Ansible SSL certificate and private key. The old paths referenced Ansible variables that were no longer used, so they have been replaced with the correct paths for the certificate and key files. This change ensures that Apache is correctly configured to use the Ansible SSL certificate and private key for secure communication.

* Refactor SSL configuration for  openssl variable  certificate generation and set ServerName to an empty string.

* Refactor VirtualHost configuration for SSL support

This commit refactors the VirtualHost configuration for SSL support. The mod_ssl module is no longer loaded with an <IfModule> directive, and the Listen directive for port 443 is removed. The SSL configuration is moved into a new <VirtualHost> block on port 443. The SSL certificate files are updated to use the correct paths. Additionally, the mod_authz_core module is used to enforce access controls in the <Directory> blocks.

* Refactor Ansible OpenSSL certificate paths

Changed the paths of OpenSSL certificates used in Ansible to ensure better organization and compliance with project standards. This commit updates the paths for OpenSSL certificates to the new designated locations.

---------

Co-authored-by: eesaanatluri <39744656+eesaanatluri@users.noreply.github.com>